+++ src/components/BusinessComponent/QueryTable/components/QueryFrom/index.tsx
@@ -1,5 +1,4 @@
 import { forwardRef, useImperativeHandle, useState } from 'react'
-import { Form, Button, Col, Row, Space } from 'antd'
 import styles from './styles.module.less'
 import { DownOutlined, UpOutlined } from '@ant-design/icons'
 import helperAntdComponet from './helper'