+++ src/pages/404/index.tsx
@@ -1,5 +1,4 @@
 import PageWarp from '@/components/BusinessComponent/PageWarp'
-import { Button, Result } from 'antd'
 
 const Component: React.FC = (): JSX.Element => {
   const navigate = useNavigate()