+++ src/pages/513/index.tsx
@@ -1,6 +1,5 @@
 import PageWarp from '@/components/BusinessComponent/PageWarp'
 import React from 'react'
-import { Button, Result } from 'antd'
 
 const Component: React.FC = (): JSX.Element => {
   const navigate = useNavigate()