+++ src/pages/QueryTable/index.tsx
@@ -1,6 +1,4 @@
 import QueryTable from '@/components/BusinessComponent/QueryTable'
-import { Space, Button } from 'antd'
-import { useMount } from 'ahooks'
 import PageWarp from '@/components/BusinessComponent/PageWarp'
 
 const dataSource = [