+++ src/pages/Todo/components/input/index.tsx
@@ -1,4 +1,3 @@
-import { Button, Input } from 'antd'
 type propsHelper = {
   addCallback: Function
 }